requirements.txt
torch>=1.9.0
torchvision>=0.10.0
pytorch-lightning>=1.5.0
torchmetrics>=0.7.0
albumentations>=1.0.0
numpy>=1.19.2
pandas>=1.3.0
Pillow>=8.3.1
PyYAML>=5.4.1
tensorboard>=2.6.0
tqdm>=4.62.2